drf/drf_testing/requirements.txt
Django==3.2.19
djangorestframework==3.14.0
